Add GetFilteredPetsWithPagination feature variation implemented via Dapper

+++ PetFamily.Backend/src/PetFamily.Application/Models/PagedList.cs
@@ -4,7 +4,7 @@ public class PagedList<T>
 {
     public required IReadOnlyList<T> Items { get; init; }
     
-    public required int TotalCount { get; init; }
+    public required long TotalCount { get; init; }
     
     public required int PageSize { get; init; }
